nits first round
