+++ tornjak-frontend/src/components/dashboard/dashboard-details-render.js
@@ -17,7 +17,7 @@ import {
     workloadSelectorInfoFunc,
     agentworkloadSelectorInfoFunc,
     clusterTypeInfoFunc,
-    clickedDashboardTabelFunc,
+    clickedDashboardTableFunc,
 } from 'redux/actions';
 
 class DashboardDetailsRender extends Component {
@@ -33,7 +33,7 @@ class DashboardDetailsRender extends Component {
         if (this.props.globalTornjakServerInfo === "" || this.props.globalServerInfo === undefined) {
             if (IsManager) {
                 if (this.props.globalServerSelected !== "") {
-                    this.props.clickedDashboardTabelFunc(params.entity + "details");
+                    this.props.clickedDashboardTableFunc(params.entity + "details");
                     this.TornjakApi.populateClustersUpdate(this.props.globalServerSelected, this.props.clustersListUpdateFunc, this.props.tornjakMessageFunc);
                     this.TornjakApi.populateAgentsUpdate(this.props.globalServerSelected, this.props.agentsListUpdateFunc, this.props.tornjakMessageFunc);
                     this.TornjakApi.populateEntriesUpdate(this.props.globalServerSelected, this.props.entriesListUpdateFunc, this.props.tornjakMessageFunc);
@@ -42,7 +42,7 @@ class DashboardDetailsRender extends Component {
                     this.TornjakApi.populateTornjakServerInfo(this.props.globalServerSelected, this.props.tornjakServerInfoUpdateFunc, this.props.tornjakMessageFunc);
                 }
             } else {
-                this.props.clickedDashboardTabelFunc(params.entity + "details");
+                this.props.clickedDashboardTableFunc(params.entity + "details");
                 this.TornjakApi.refreshLocalSelectorsState(this.props.agentworkloadSelectorInfoFunc);
                 this.TornjakApi.populateLocalClustersUpdate(this.props.clustersListUpdateFunc, this.props.tornjakMessageFunc);
                 this.TornjakApi.populateLocalAgentsUpdate(this.props.agentsListUpdateFunc, this.props.tornjakMessageFunc);
@@ -75,5 +75,5 @@ const mapStateToProps = (state) => ({
 
 export default connect(
     mapStateToProps,
-    { serverSelectedFunc, clustersListUpdateFunc, agentsListUpdateFunc, entriesListUpdateFunc, tornjakServerInfoUpdateFunc, serverInfoUpdateFunc, selectorInfoFunc, tornjakMessageFunc, workloadSelectorInfoFunc, agentworkloadSelectorInfoFunc, clusterTypeInfoFunc, clickedDashboardTabelFunc }
+    { serverSelectedFunc, clustersListUpdateFunc, agentsListUpdateFunc, entriesListUpdateFunc, tornjakServerInfoUpdateFunc, serverInfoUpdateFunc, selectorInfoFunc, tornjakMessageFunc, workloadSelectorInfoFunc, agentworkloadSelectorInfoFunc, clusterTypeInfoFunc, clickedDashboardTableFunc }
 )(DashboardDetailsRender)
